+++ src/gui/shared/head_morph.rs
@@ -42,7 +42,6 @@ impl Component for HeadMorph {
     type Properties = Props;
 
     fn create(ctx: &Context<Self>) -> Self {
-        // TODO: Import gibbed head morph
         let save_handler = SaveHandler::bridge(ctx.link().callback(|response| match response {
             Response::HeadMorphImported(head_morph) => Msg::HeadMorphImported(head_morph),
             Response::HeadMorphExported => Msg::HeadMorphExported,